Refactor code to use codeowners.go more
